refactor autogenra -> autogenstudio
