--- samples/apps/autogen-assistant/autogenra/__init__.py
@@ -1,3 +1,0 @@
-from .autogenflow import *
-from .autogenchat import *
-from .datamodel import *
