+++ samples/apps/autogen-studio/MANIFEST.in
