+++ samples/apps/autogen-studio/autogenstudio/datamodel.py
