+++ samples/apps/autogen-studio/autogenstudio/web/skills/global/generate_images.py
