+++ samples/apps/autogen-studio/frontend/.gitignore
