+++ samples/apps/autogen-studio/frontend/gatsby-browser.js
