+++ samples/apps/autogen-studio/frontend/gatsby-ssr.tsx
