+++ samples/apps/autogen-studio/frontend/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "AutoGen_Assistant",
+  "name": "AutoGen_Studio",
   "version": "1.0.0",
   "private": true,
-  "description": "AutoGen Assistant - Build LLM Enabled Agents",
+  "description": "AutoGen Studio - Build LLM Enabled Agents",
   "author": "SPIRAL Team",
   "keywords": [
     "gatsby"
@@ -10,7 +10,7 @@
   "scripts": {
     "develop": "gatsby clean && gatsby develop",
     "start": "gatsby clean && gatsby develop",
-    "build": "gatsby clean && rm -rf ../autogenra/web/ui  && PREFIX_PATH_VALUE='' gatsby build --prefix-paths && cp -r  public/ ../autogenra/web/ui",
+    "build": "gatsby clean && rm -rf ../autogenstudio/web/ui  && PREFIX_PATH_VALUE='' gatsby build --prefix-paths && cp -r  public/ ../autogenstudio/web/ui",
     "serve": "gatsby serve",
     "clean": "gatsby clean",
     "typecheck": "tsc --noEmit"
