+++ samples/apps/autogen-studio/frontend/src/components/footer.tsx
