+++ samples/apps/autogen-studio/frontend/src/components/header.tsx
