+++ samples/apps/autogen-studio/frontend/src/components/icons.tsx
