+++ samples/apps/autogen-studio/frontend/src/components/layout.tsx
