+++ samples/apps/autogen-studio/frontend/src/components/views/gallery/gallery.tsx
