+++ samples/apps/autogen-studio/frontend/src/components/views/ra/sidebar.tsx
