+++ samples/apps/autogen-studio/frontend/src/images/icon.png
