+++ samples/apps/autogen-studio/frontend/src/pages/gallery/index.tsx
