+++ samples/apps/autogen-studio/frontend/src/pages/index.tsx
