+++ samples/apps/autogen-studio/frontend/src/styles/global.css
