+++ samples/apps/autogen-studio/frontend/tailwind.config.js
