+++ samples/apps/autogen-studio/frontend/tsconfig.json
